Abstract the sideOf and uuid functions.
